requirements.txt
libmr==0.1.9
matplotlib==3.3.2
numpy==1.19.4
Pillow==10.1.0
pretrainedmodels==0.7.4
scikit_learn==1.3.2
scipy==1.5.2
timm==0.8.0.dev0
torch==1.9.0
torchvision==0.10.0
